Shift around util functions
